Feat: Introduce dynamic color scheme updates in AppTheme
This commit introduces the ability to dynamically update the color scheme in `AppTheme`.
It defines a `MaterialColor` interface and an `AppThemeColorScheme` class that allows modifying individual colors of the color scheme.
A `LocalMaterialColor` composition local is introduced to provide this capability to composables down the tree.
`MaterialColorScheme` sealed interface is also introduced to support easy access and modification of material color.
A button is added in the `App` composable to demonstrate the color update.