Feat: Add Button component implementations and component details screen
This commit introduces several button implementations: Filled, Text, Elevated, Tonal, Outlined, and Icon.
It also introduces a new `ComponentDetails` screen that is a generic screen that can show a list of Material components.
Additionally a `ComponentItem` and `Component` were introduced to be able to show the list of the component details.
A new dependency `kotlinx.collections.immutable` was introduced to support the usage of immutable lists.